+++ scripts/setup.sql
@@ -1,0 +1,15 @@
+-- This script sets up the initial database schema and test data for the CDC-Lite project
+-- It creates a sample users table and populates it with initial test data
+
+-- Create a sample users table with basic user information
+CREATE TABLE users (
+    id INTEGER PRIMARY KEY,
+    name VARCHAR,
+    email VARCHAR,
+    created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
+);
+
+-- Insert initial test data
+INSERT INTO users (id, name, email) VALUES 
+    (1, 'John Doe', 'john@example.com'),
+    (2, 'Jane Smith', 'jane@example.com');
