--- setup.sql
@@ -1,12 +1,0 @@
--- Create a sample users table
-CREATE TABLE users (
-    id INTEGER PRIMARY KEY,
-    name VARCHAR,
-    email VARCHAR,
-    created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
-);
-
--- Insert initial data
-INSERT INTO users (id, name, email) VALUES 
-    (1, 'John Doe', 'john@example.com'),
-    (2, 'Jane Smith', 'jane@example.com');
